refactor rocm files
